requirements.txt
asgiref==3.11.1
Django==6.0.3
psycopg2-binary==2.9.11
python-dotenv==1.0.1
sqlparse==0.5.5
djangorestframework==3.17.1
